3. Generative AI Live/10.1. Project-III/7. Simple Retriever App/requirements.txt
langchain==0.2.11
langchain-openai==0.1.20
python-dotenv==1.0.1
langchain-community==0.2.10
pypdf==4.3.1
langchain-chroma==0.1.2